pyproject.toml
[project]
name = "spotify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.3",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "spotipy>=2.25.1",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
